[#15] Fix pagination bug in harvesters
To check if the remote server supported pagination we were only checking
whether the content from a request was the same as the previous one.
This is quite fragile as some fields might get updated on each request,
eg modified date for realtime data.

We are now checking if the guids from a request are the same as the
previous ones, which should be more reliable.