[#28] Add RDF endpoint for whole catalog
RDF profiles can implement a `graph_from_catalog` method to generate a
dcat:Catalog graph.

RDF Parser has a `serialize_catalog` method to provide an RDF
serialization for the whole catalog. A list of datasets can be provided to add
them to the serialization. This endpoint uses the same pagination as the
odl JSON endpoint (`page` and `modified_since` parameters). The page
size defaults to 100 and can be configured server side.

The serialization is accessible at `http://<catalog>/catalog.<format>`