[#28] Add functions to generate URIs if not provided
For the catalog, the recommended way is to define the
`ckanext.dcat.base_uri` config option, if not `ckan.site_url` is used,
and if this is not present, one is created using the app id, but this
will cause a warning.

For datasets, if a uri is not present in the metadata, one will be
constructed with the catalog URI and the dataset id.

For resources, if a uri is not present in the metadata, one will be
constructed with the catalog URI, dataset id and resource id.

For publishers, if a uri is not present in the metadata, one will be
constructed with the catalog URI and the org id