[#28] Better handling of format field when parsing
Media type and format label are extracted and assigned to the relevant
resource fields. Values are extracted from dcat:mediaType and
dct:format. For the latter, the following form is also supported, as it
is really common:

<dct:format>
    <dct:IMT rdf:value="text/html" rdfs:label="HTML"/>
</dct:format>

By default the formats are normalized in the same way that when creating
a resource, using the resource config on ckan core.