[#75] Add missing dataset fields
The SchemaOrg schema provides a number of methods that could be
overriden by subclasses if they need a slightly different mapping to
schema.org (e.g. from one of the DCAT AP standards).

With those methods in place it should be relatively easy to implement a
schema.org mapping for different standards.