Add exception handling for import_stage.
Without exception handling in place it will cause the fetch process to die so it will have to be restarted. Also the user would be unaware of any errors as they are not caught and saved to the error log.
In addition to this if the process dies too often in a short space of time the process might not restart it due to too many failures, which will lead to no jobs completing as the process is not running. Also although the number of retries before the process dies can be increased to make it more resilient in practise this will slow things down so it is better to log and handle them.