Output geometries as WKT only by default
The Shacl validation expects only one item of LOCN.geometry, DCAT.bbox
or DCAT.centroid. Up until now we were adding two triples, one for
GeoJOSN and one for WKT. We'll default to WKT from now on as this is
what GeoDCAT-AP requires (or GML yuk...), but sites that for some reason
require GeoJSON (or both) can use the
`ckanext.dcat.output_spatial_format` to choose which format to use.