Avoid duplicated dates in DCAT-AP 2 profile
The profile for DCAT-AP 1 stored triples using schema:startDate/endDate
but the namespace was updated to dcat:startDate/endDate in DCAT-AP 2.
As the dcat-ap 2 profile calls the version 1 profile, the two sets of
dates were added together. This makes fail the shapes recommended Shacl
tests and it is incorrect anyway, so we now remove the schema-namespaced
triples if present when using the dcat-ap 2 profile.