Update permissions in tests container runner image
After changes in the CKAN Docker images (ckan/ckan-docker-base#80) the
images run in the test workflows does not use root by default anymore.
This is not supported in GitHub Actions so we need to pass `--user root`
option to avoid failures. To catch any permissions related issues, we
run the `ckan` and `pytest` commands with `sudo -u ckan`