Update tests for changes in date parsing in rdflib
Rdflib 7.1 introduced changes in [date parsing][1] that made some base
profile tests fail. Basically the previous rdflib versions incomplete
dates like

    <time:inXSDDateTimeStamp
         rdf:datatype="http://www.w3.org/2001/XMLSchema#date">
            1904
    </time:inXSDDateTimeStamp>

were expanded to `1904-01-01`. Of course this is not a valid date and
should be expressed using `gYear`:

    <time:inXSDDateTimeStamp
         rdf:datatype="http://www.w3.org/2001/XMLSchema#gYear">
            1904
    </time:inXSDDateTimeStamp>

and we should be expecting `1904`.
This should play nice with the time properties we are generating in CKAN as
they already handle automatically `gYear`, `gYearMonth`, `date` and
`dateTime`.
Sites importing external DCAT representations that use the wrong
encoding might need to check their parsers.

[1] https://github.com/RDFLib/rdflib/pull/2929