Remove default `@language` property from Croissant context
This makes the output much clearer by removing all `@value`
language tags, and besides we can not assume what the language of the
metadata in different CKAN sites will be so it's better to let users
create their custom profiles with whatever makes sense.

This will emit a warning when validating the Croissant output with the
mlcroissant package (e.g. in the test) but I think overall it's worth
doing.