Provide default language in Croissant JSON-LD context
This was to fixed failures caused by the new version of the croissant
package (introduced by https://github.com/mlcommons/croissant/pull/932),
but it seems like a good thing to do anyway.

Sadly this means that the output generated from rdflib changes, from
this:

```
"description": "Test description"
```

to this:

```
"description": {
    "@value": "Test description"
}
```

While this is a bit less human readable is perfectly valid JSON-LD, and
it will probably align better with multilingual metadata.